Fix socket closing.
The C call shutdown( fd, SHUT_WR ) was executed when closing the
socket while it must be called when the socket output port is
closed.

This confusion raised errors where the output port was closed, then
potentially reused by another socket and then shutdown when the
initial socket was closed. This was the reason for hop broadcasting
and service calls not working properly anymore.