Changes to support using gcc's link time optimization
The majority of the changes are minor discrepancies between types
declared in the bigloo runtime c files and their uses in bigloo
source file extern clauses.

The second type of changes were modifications to either prevent the
inlining of procedures essential to proper operation of callcc or to
address optimizations that prevented the proper operation of
callcc. For the latter, this specifically referse to the use of alloca
in callcc_restore_stack to grow the stack due to the recursive call
being turned into a goto.