Stack overflow detection
This is not complete yet as it does not handle correctly
situations where the stack overflow occurs in the middle of a
GC allocation or reclaim. If that happens the GC data structures
might be corrupted and the recovery is not reliable. It is unclear
if this could be fixed.