Add suport for adding a prefix to module exports
extend the module system to support prefixing exported symbols with a
user-provided symbol. For example, the following prefixes all of the
exports from the test module with t-.

(import (:prefix t- test))

You can also prefix an explicitly specified set of symbols:

(import (:prefix t- doit fizzbuzz test))

In this case, doit and fizzbuzz are available as t-doit and
t-fizzbuzz.

The pre-exisiting alias functionality is still available and takes
precendence.

(import (:prfix t- (my-doit doit) fizzbuzz test))

In the above, doit is available as my-doit and fizzbuzz as t-fizzbuzz