pthread modifications and corrections
Use pthread_cancel and the PTHREAD_CANCEL_DEFERRED cancel type to achieve safer, but
not safe, thread-terminate! behavior.

Correct the semantics of the java implementation of condition
variables. When condition-variable-wait! returns, regardless of
success or failure, the associated mutex should be locked.

thread-sleep was not updated when the standard bigloo sleep procedure
was moved to microseconds. Make sure we pass microseconds not milliseconds.