fix profile build in jvm backend
Bigloo reported a problem finding clinit when compiling bigloo
classes. This was due to clinit being declared for both bigloo modules
and bigloo classes but only actually being defined for bigloo
modules. Additionally, made a small correction to the clinit generated
for profiling when no existing clinit definition was available for
modification; a return restruction was appended to the clinit
profiling logic.