Alter get-syntax-expander to use hygiene-value* instead of hygien-value
hygien-value* strips of all copies of hygiene-mark before returning
the residual id. hygiene-value* only replaces hygiene-value in
get-syntax-expander. Other occurrences still usy hygiene-value which
only strips of the first hygiene-mark prefix. This is necessary to
maintain macro hygiene. With this change, we are able to properly
expand the CPS syntax-rules macros used in SRFI-42