a couple jvm backend fixes
The first is a correction to rgc_buffer_eof_p which brings it in
alignment with the c implementation and corrects a corner case where
we were at the end of file but not correctly reporting so. The second
adds missing configuration options (STRICT_STDCC and
BGL_HAVE_GETRLIMIT) which were omitted for the jvm backend. Finally, a
small typo correction was made to bglgst.c to permit compilation.