Correct miscompilation of boxes with -fstackable
We add a stackable walk-method for make-box that properly takes into
account whether or not the corresponding var is escaping. Without this
change, Bigloo would believe it is correct to stack allocate an
escaping variable bound to a make-box, resulting in interpreting an
arbitrary stack location as a Bigloo object. Below is a simple
procedure that was being handled incorrectly when compiled with
-call/cc and -fstackable (or -O6 which enables -fstackable). -call/cc
resulted in x being boxed and -fstackable was mistakenly stack
allocating it which resulted in errors when accessing x in the closure
bound to doit.

(define doit
   (let ((x -1))
      (lambda (e)
         (set! x (+ x 1))
         (string-append "my string"
            (number->string x)
            e))))