Changes to the Bigloo C runtime to support C23 and GCC 15
Most changes are due to C23 which GCC 15 defaults to changing the
meaning of function declarations without parameters (e.g., int foo()). In C17 and
earlier versions of C, this indicates that there is no information
about the number and type of the arguments required by the function,
requiring the developer to know the correct arguments but not
restricting what arguments may be passed to the function. In C23, a
function with empty parameters indicates a function takes no
arguments. The Bigloo C runtime relied on this in a number of
places. To compile under C23 and GCC 15, we replaced such
constructs. This largely entailed the addition of proper function
declarations and updating autoconf scripts that relied on the
feature. It also means that C23 support requires BGL_STRICT_STDC be
true.

The changes have been successfully tested with GCC 15, GCC 14, and
clang 19. Only GCC 15 defaults to the C23 standard.